refactor folder structure, split simplelist into smaller connected components
