+++ src/client/components/top/TopBar.component.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import {connect} from 'react-redux';
 import Link from '../general/Link.component';
-import ShortListDropDown from '../list/ShortListDropDown.container';
-import ListOverviewDropDown from '../list/ListOverviewDropDown.container';
+import ShortListDropDown from '../list/shortlist/ShortListDropDown.container';
+import ListOverviewDropDown from '../list/overview/ListOverviewDropDown.container';
 import ProfileImage from '../general/ProfileImage.component';
 import Icon from '../base/Icon';
 import SearchBar from '../filter/SearchBar.component';
